added new method send message on emulator

+++ UBot/Whatsapp/Client.cs
@@ -3,6 +3,7 @@ using System.Runtime.InteropServices;
 using UBot.Whatsapp.Web;
 using MemuLib.Core;
 using System.Reflection.Metadata;
+using static Microsoft.Maui.ApplicationModel.Permissions;
 
 namespace UBot.Whatsapp;
 
@@ -49,6 +50,9 @@ public class Client
 
     public async Task UpdateData(bool PullAccount)
     {
+        if (!File.Exists($@"{Account}\Data.json"))
+            return; 
+
         await File.WriteAllTextAsync($@"{Account}\Data.json", JsonConvert.SerializeObject(AccountData, Formatting.Indented));
         
         if (!PullAccount)
@@ -78,67 +82,69 @@ public class Client
             await Mem.RunApk(PackageName);
 
             await Task.Delay(1_500);
+            var lastCalled = false;
+            while (true)
+            {
+                var dump = await Mem.DumpScreen();
 
-            var dump = await Mem.DumpScreen();
+                if (!await Mem.ExistsElement("text=\"ПРОПУСТИТЬ\"", dump))
+                    goto s1;
 
-            if (!await Mem.ExistsElement("text=\"ПРОПУСТИТЬ\"", dump))
-                goto s1;
+                await Mem.Click("text=\"ПРОПУСТИТЬ\"");
+                await Task.Delay(1_000);
+                await Mem.StopApk(PackageName);
+                await Mem.RunApk(PackageName);
+                await Task.Delay(1_000);
+                lastCalled = true;
+                continue;
+            s1:
+                if (!await Mem.ExistsElement($"resource-id=\"{PackageName}:id/registration_name\"", !lastCalled ? dump : null))
+                    goto s2;
 
-            await Mem.Click("text=\"ПРОПУСТИТЬ\"");
-            await Task.Delay(1_000);
-            await Mem.StopApk(PackageName);
-            await Mem.RunApk(PackageName);
-            await Task.Delay(1_000);
-            dump = await Mem.DumpScreen();
+                await Mem.Input("text=\"Введите своё имя\"", name.Replace(' ', 'I'), !lastCalled ? dump : null);
+                await Mem.Click("text=\"ДАЛЕЕ\"", !lastCalled ? dump : null);
+                await Task.Delay(1_000);
+                await Mem.StopApk(PackageName);
+                await Mem.RunApk(PackageName);
+                await Task.Delay(1_000);
+                lastCalled = true;
+                continue;
+            s2:
+                if (!await Mem.ExistsElement($"resource-id=\"{PackageName}:id/code\"", !lastCalled ? dump : null))
+                    goto s3;
 
-        s1:
-            if (!await Mem.ExistsElement($"resource-id=\"{PackageName}:id/registration_name\"", dump))
-                goto s2;
+                await Mem.Input($"resource-id=\"{PackageName}:id/code\"", Globals.Setup.PinCodeAccount.ToString(), !lastCalled ? dump : null);
+                await Mem.StopApk(PackageName);
+                await Mem.RunApk(PackageName);
+                await Task.Delay(1_000);
+                lastCalled = true;
+                continue;
+            s3:
+                if (!await Mem.ExistsElement("text=\"НЕ СЕЙЧАС\"", !lastCalled ? dump : null))
+                    goto s4;
 
-            await Mem.Input("text=\"Введите своё имя\"", name.Replace(' ', 'I'), dump);
-            await Mem.Click("text=\"ДАЛЕЕ\"", dump);
-            await Task.Delay(1_000);
-            await Mem.StopApk(PackageName);
-            await Mem.RunApk(PackageName);
-            await Task.Delay(1_000);
-            dump = await Mem.DumpScreen();
+                await Mem.Click("text=\"НЕ СЕЙЧАС\"", !lastCalled ? dump : null);
+                await Task.Delay(1_000);
+                await Mem.StopApk(PackageName);
+                await Mem.RunApk(PackageName);
+                await Task.Delay(1_000);
+                lastCalled = true;
+                continue;
+            s4:
+                if (!await Mem.ExistsElement("text=\"Выберите частоту резервного копирования\"", !lastCalled ? dump : null))
+                    return true;
 
-        s2:
-            if (!await Mem.ExistsElement($"resource-id=\"{PackageName}:id/code\"", dump))
-                goto s3;
-
-            await Mem.Input($"resource-id=\"{PackageName}:id/code\"", Globals.Setup.PinCodeAccount.ToString(), dump);
-            await Mem.StopApk(PackageName);
-            await Mem.RunApk(PackageName);
-            await Task.Delay(1_500);
-            dump = await Mem.DumpScreen();
-
-        s3:
-            if (!await Mem.ExistsElement("text=\"НЕ СЕЙЧАС\"", dump))
-                goto s4;
-
-            await Mem.Click("text=\"НЕ СЕЙЧАС\"", dump);
-            await Task.Delay(1_000);
-            await Mem.StopApk(PackageName);
-            await Mem.RunApk(PackageName);
-            await Task.Delay(1_000);
-            dump = await Mem.DumpScreen();
-
-        s4:
-            if (!await Mem.ExistsElement("text=\"Выберите частоту резервного копирования\"", dump))
-                return true;
-
-            //await Mem.Click(374, 513);//Выберите частоту резервного копирования
-            //await Mem.Click(414, 846);//text=\"Никогда\"
-            //await Mem.Click(616, 1238);//text=\"ГОТОВО\"
-            await Mem.Click("text=\"Выберите частоту резервного копирования\"", dump);
-            await Mem.Click("text=\"Никогда\"", dump);
-            await Mem.Click("text=\"ГОТОВО\"", dump);
-            await Task.Delay(1_000);
-            await Mem.StopApk(PackageName);
-            await Mem.RunApk(PackageName);
-
-            return true;
+                //await Mem.Click(374, 513);//Выберите частоту резервного копирования
+                //await Mem.Click(414, 846);//text=\"Никогда\"
+                //await Mem.Click(616, 1238);//text=\"ГОТОВО\"
+                await Mem.Click("text=\"Выберите частоту резервного копирования\"", !lastCalled ? dump : null);
+                await Mem.Click("text=\"Никогда\"", !lastCalled ? dump : null);
+                await Mem.Click("text=\"ГОТОВО\"", !lastCalled ? dump : null);
+                await Task.Delay(1_000);
+                await Mem.StopApk(PackageName);
+                await Mem.RunApk(PackageName);
+                continue;
+            }
         }
         catch (Exception ex)
         {
@@ -150,7 +156,14 @@ public class Client
     public async Task<bool> ImportContacts(string path)
     {
         await Mem.ClearContacts();
-        await Mem.ImportContacts(path);
+
+        var rndName = $"{new Random().Next(1_000, 1_000_000)}_contacts.vcf";
+
+        var file = new FileInfo(path);
+        file.CopyTo(@$"{Globals.Setup.PathToDownloadsMemu}\{rndName}");
+
+        await MemuCmd.ExecMemuc(
+            $"-i {Mem.Index} execcmd am start -t \"text/x-vcard\" -d \"file:///storage/emulated/0/Download/{rndName}\" -a android.intent.action.VIEW cz.psencik.com.android.contacts");//com.android.contacts");
 
         var dump = await Mem.DumpScreen();
 
@@ -165,26 +178,19 @@ public class Client
         return true;
     }
 
-    public async Task<bool> SendMessage(string to, string text, FileInfo image = null)
+    /*public async Task<bool> SendMessage(string toPhone, string text)
     {
+        var to = (toPhone[0] == '+') ? toPhone : $"+{toPhone}";
         //await Mem.Shell($@"am start -a android.intent.action.VIEW -d https://wa.me/{to}/?text={Uri.EscapeDataString(text)}");
-        var command = new FileInfo($@"{Globals.TempDirectory.FullName}\{to}.sh");
+        var command = new FileInfo($@"{Globals.Setup.PathToDownloadsMemu}\{to}.sh");
         var isSended = false;
-        var commandImage = string.Empty;
-
-        if (image != null && image.Exists)
-        {
-            if (!File.Exists($@"{Globals.Setup.PathToDownloadsMemu}\{image.Name}"))
-                new FileInfo(image.FullName).CopyTo($@"{Globals.Setup.PathToDownloadsMemu}\{image.Name}");
-
-            commandImage = $" --eu android.intent.extra.STREAM file:///storage/emulated/0/Download/{image.Name}";//Тут добавлен пробел для синхроности, я мудак идите нахуй
-        }
 
         await File.WriteAllTextAsync(command.FullName,
-            $"NL=$'\\n' ; am start -a android.intent.action.SEND --es android.intent.extra.TEXT \"{text.Replace("\r", "${NL}").Replace("\n", "${NL}")}\" -t text/plain -e jid '{to}@s.whatsapp.net'{commandImage} -p com.whatsapp.w4b");//$"am start -a android.intent.action.VIEW -d https://wa.me/{to}/?text={Uri.EscapeDataString(text)} {PackageName}");//mb fix
+            $"am start -a android.intent.action.VIEW -d https://wa.me/{to}/?text={Uri.EscapeDataString(text)} {PackageName}");//mb fix
 
-        await Mem.Push(command.FullName, "/data/local/tmp");
-        await Mem.Shell($@"sh /data/local/tmp/{to}.sh");
+        Log.Write(await Mem.Shell($@"sh /storage/emulated/0/Download/{command.Name}"));
+        //await Mem.Push(command.FullName, "/data/local/tmp");
+        //await Mem.Shell($@"sh /data/local/tmp/{to}.sh");
 
         for (var i = 0; i < 3; i++)
         {
@@ -197,24 +203,71 @@ public class Client
                 if (await Mem.ExistsElement("text=\"OK\"", dump, false))
                     await Mem.Click("text=\"OK\"", dump);
 
-                /*var document = await Mem.DumpScreen();
-                if (await Mem.ExistsElement("text=\"ОК\"", document))
-                    await Mem.Click("text=\"ОК\"", document);
-
-                if (await Mem.ExistsElement("text=\"OK\"", document))
-                    await Mem.Click("text=\"OK\"", document);*/
-
-                await Task.Delay(500);
+                await Task.Delay(1_500);
                 continue;
             }
 
             await Mem.Click("content-desc=\"Отправить\"", dump);
             isSended = true;
-            AccountData.MessageHistory[to] = DateTime.Now;
+            //AccountData.MessageHistory[to] = DateTime.Now;
             break;
         }
 
-        await Mem.Shell($"rm /data/local/tmp/{to}.sh");
+        //await Mem.Shell($"rm /data/local/tmp/{to}.sh");
+        File.Delete(command.FullName);
+
+        return isSended;
+    }*/
+
+    /*
+     * Нужно сделать хотфикс на обновление контактов
+     * Чтобы работал метод
+     */
+    public async Task<bool> SendMessage(string contact, string text, FileInfo image = null)
+    {
+        var to = contact.Replace("+", "");
+        //await Mem.Shell($@"am start -a android.intent.action.VIEW -d https://wa.me/{to}/?text={Uri.EscapeDataString(text)}");
+        var command = new FileInfo($@"{Globals.Setup.PathToDownloadsMemu}\{to}.sh");
+        var isSended = false;
+        var commandImage = string.Empty;
+
+        if (image != null && image.Exists)
+        {
+            if (!File.Exists($@"{Globals.Setup.PathToDownloadsMemu}\{image.Name}"))
+                new FileInfo(image.FullName).CopyTo($@"{Globals.Setup.PathToDownloadsMemu}\{image.Name}");
+
+            commandImage = $" --eu android.intent.extra.STREAM file:///storage/emulated/0/Download/{image.Name}";//Тут добавлен пробел для синхроности, я мудак идите нахуй
+        }
+        //execcmd am start -e jid 79772801086@s.whatsapp.net com.whatsapp.w4b/com.whatsapp.Conversation
+        await File.WriteAllTextAsync(command.FullName,
+        //NL=$'\n' ; am start -a android.intent.action.SEND --es android.intent.extra.TEXT "Hello ${NL} World" -t text/plain -e jid '79772801086@s.whatsapp.net' --eu android.intent.extra.STREAM file:///storage/emulated/0/Download/1.jpg com.whatsapp.w4b/com.whatsapp.Conversation 
+        //NL=$'\\n' ; am start -a android.intent.action.SEND --es android.intent.extra.TEXT \"{text.Replace("\r", "${NL}").Replace("\n", "${NL}")}\" -t text/plain -e jid '{to}@s.whatsapp.net'{commandImage} com.whatsapp.w4b/com.whatsapp.Conversation 
+        $"NL=$'\\n' ; am start -a android.intent.action.SEND --es android.intent.extra.TEXT \"{text.Replace("\r", "${NL}").Replace("\n", "${NL}")}\" -t text/plain -e jid '{to}@s.whatsapp.net'{commandImage} com.whatsapp.w4b");
+
+        Log.Write(await Mem.Shell($@"sh /storage/emulated/0/Download/{command.Name}"));
+
+        for (var i = 0; i < 3; i++)
+        {
+            var dump = await Mem.DumpScreen();
+            if (!await Mem.ExistsElement("content-desc=\"Отправить\"", dump, false))
+            {
+                if (await Mem.ExistsElement("text=\"ОК\"", dump, false))
+                    await Mem.Click("text=\"ОК\"", dump);
+
+                if (await Mem.ExistsElement("text=\"OK\"", dump, false))
+                    await Mem.Click("text=\"OK\"", dump);
+
+                await Task.Delay(1_000);
+                continue;
+            }
+
+            await Mem.Click("content-desc=\"Отправить\"", dump);
+            isSended = true;
+            //AccountData.MessageHistory[to] = DateTime.Now;
+            break;
+        }
+
+        //Log.Write(await Mem.Shell($"rm /data/local/tmp/{to}.sh"));
         File.Delete(command.FullName);
 
         return isSended;
